Fix 32 bit build

+++ src/debug.c
@@ -1104,7 +1104,7 @@ void _serverAssertPrintClientInfo(const client *c) {
     bugReportStart();
     serverLog(LL_WARNING, "=== ASSERTION FAILED CLIENT CONTEXT ===");
     for (size_t i = 0; i < sizeof(c->raw_flag) / sizeof(c->raw_flag[0]); i++) {
-        serverLog(LL_WARNING, "client->flags[%ld] = %llu", i, (unsigned long long)c->raw_flag[i]);
+        serverLog(LL_WARNING, "client->flags[%lu] = %llu", (unsigned long)i, (unsigned long long)c->raw_flag[i]);
     }
     serverLog(LL_WARNING, "client->conn = %s", connGetInfo(c->conn, conninfo, sizeof(conninfo)));
     serverLog(LL_WARNING, "client->argc = %d", c->argc);
